ECPREOPSTK-105-Validation script for CreateConfiguration procedure
Added a validation script for CreateConfiguration to be used with
ValidateFormalParameters API by Dynamic Environments.

o Added createConfiguration.groovy script that will be used by the
ValidateFormalParameters API in order to validate the
CreateConfiguration parameters.
o Added optional tenantId parameter to CreateConfiguration. It is
required to validate the service end-point URLs since the service
catalog is associated with the tenant/project.
o Updated the createConfigForm.xml to mark the following parameters for
server-side validation. This information is used by the UI to trigger
the server-side call for the marked parameters:
- Identity URL
- Credential
- Tenant id
- Compute service URL
- Block service URL
- Image service URL
- Orchestration service URL
o Updated build.gradle to support running groovy tests