Updated the validation and dynamic options script for credential support
o Updated the validation and dynamic options scripts to use the
credential argument value for accessing userName/password values. The
credential values are either provided by the user or the API will
retrieve the right attached credential value and pass to the script.
o Updated the dynamic options scripts to use the configurationParameters
argument to access the associated configuration values.