Added properties for credentials and configurations retrieval support
Added following properties for credentials and configurations retrieval
support for the formal parameter APIs.

1) ec_form/configurationParameterRef: Procedure property for the formal
parameter name in the procedure that refers to the plugin configuration
that the procedure uses. For _DeployDE procedure, this property is set
to 'connection_config'.
2) ec_form/usesConfigurationNameAsCredential: Procedure property for
whether the configuration name (referred to in the previous property)
should be used to retrieve the attached credential for the procedure.
For _DeployDE procedure, this property is set to '1' so that the
attached credential can be retrieved using the value set in the
'config_connection' parameter.
3) ec_configurations: Plugin project property for the path where the
configurations are stored relative to the project. For EC-OpenStack
plugin, this property is set to 'openstack_cfgs'.

Also renamed the 'form/validation' and 'form/parameterOptions' property
sheets to 'ec_form/validation' and 'ec_form/parameterOptions'
respectively.