Invalid password error message
The error from OpenStack is unpredictable when only the password is
invalid. Sometimes it returns "AuthenticationException: Unauthorized"
and other times it returns "ServerResponseException: Internal server
error". I think this is done intentionally to make it difficult for
someone to hack as they would not be able to tell whether the user id or
the password is invalid. But that means that it is also difficult for us
to tell whether the error is due to an invalid password or for some
other reason when the error is "ServerResponseException: Internal server
error".
So, addressing this case by reporting this error as a general
authentication failure against both user name and password with the
general failure to authenticate error message.